requirements.txt
opencv-python
loguru
monai==1.0.1
pyrender==0.1.33
smplx==0.1.28
scikit-learn
scikit-image
jpeg4py
matplotlib
pandas
configargparse
flatten_dict
chumpy
numpy==1.23.1
yacs
networkx>=2.6
charset-normalizer==3.1.0
tensorboard
sentence-transformers>=2.2.2
peft
spacepy
h5py